fix: rename serverless
