+++ src/stacks/api/report/express-app.ts
@@ -8,7 +8,7 @@ import {
 } from '@aws-sdk/client-dynamodb';
 import { marshall, unmarshall } from '@aws-sdk/util-dynamodb';
 import express from 'express';
-import { AuthorizerContext, IReport, QueryStringParameters, REPORT_TABLE_NAME } from './index';
+import { AuthorizerContext, IReport, QueryStringParameters, REPORT_TABLE_NAME } from './serverless';
 
 declare global {
   namespace Express {
